requirements.txt
adal==1.2.1
aniso8601==8.1.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.4
Click==7.0
colorama==0.4.4
cryptography==2.8
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
filelock==3.0.12
isodate==0.6.0
isort==5.6.4
itsdangerous==1.1.0
jsonpickle==1.2
lazy-object-proxy==1.4.3
Mako==1.0.4
MarkupSafe==1.1.1
mccabe==0.6.1
msal==1.2.0
msrest==0.6.10
mysql-connector-python==8.0.21
mysqlclient==2.0.1
oauthlib==3.1.0
pbr==3.1.1
pipenv==2020.11.15
protobuf==3.14.0
py==1.4.31
py2neo==3.1.2
pycparser==2.14
PyJWT==2.0.1
pylint==2.6.0
PyMySQL==0.10.1
pypinyin==0.22.0
python-dateutil==2.8.1
python-editor==1.0.1
pytz==2020.4
PyYAML==5.3.1
requests==2.12.1
requests-oauthlib==1.3.0
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
six==1.12.0
SQLAlchemy==1.0.14
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
virtualenv==20.4.0
virtualenv-clone==0.5.4
Werkzeug==0.15.4
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.0.0
xlwt==1.2.0
yaml-1.3==0.1.0
